SRAM stuff
When looking at the pokegear map (specifically, when decompressing the graphic in TownMapGFX) something writes to SRAM Bank 4. I can't be bothered to find out what, so instead I just put the third group of boxes in SRAM Bank 5.

This function also introduces an infuriating bug when emulating in SameBoy, in which wCurDay is decremented by 24. Who knows what's going on there. Opening the pokedex does the same thing.